chore: Avoid using git submodule for new project template (#119)
[`cargo publish` does not include the new project template as it has
`Cargo.toml`](https://doc.rust-lang.org/cargo/reference/manifest.html#the-exclude-and-include-fields),
so we need to rename it to `Cargo.toml.template` or remove it, and
having an external repo without Cargo.toml defeats the purpose. Thus, I
merged the code into this repo and we will need to set up CI to test
that template.